Add color-coding to street blocks
Street blocks are now styled depending on the value of
PARKING_CATEGORY in the entries attributes. Symbol holds a color and
a PathEffect so paths can be dashed, for example.

LayerTileProvider now has access to the Moshi instance through the
SdotRestClient singleton so it can parse PARKING_CATEGORY correctly.
This also takes the MapService out of the constructor.